2.13.2 fixes.

+++ src/main/scala/evolve/core/Evolver.scala
@@ -93,7 +93,7 @@ object Evolver {
     // returns the best child not worse than the parent
     if(optimise) {
       (( program, programScore ) +: childResults)
-        .map( a => a.copy( _2 = (a._2 + a._1.cost) * ( if( optimiseForPipeline ) a._1.maxPipelineLength else 1.0 ) ) )
+        .map( a => a.copy( _2 = (a._2 + a._1.cost) * ( if( optimiseForPipeline ) a._1.maxPipelineLength.toDouble else 1.0 ) ) )
         .reduceOption[(Program, Double)] {
         case (a, b) => if( a._2 < b._2 ) a else b
       }.map( _._1 ).filterNot( _ == program )
